+++ src/directives/bounds.js
@@ -1,40 +1,56 @@
-angular.module("leaflet-directive").directive('bounds', function ($log) {
+angular.module("leaflet-directive").directive('bounds', function ($log, leafletHelpers) {
     return {
         restrict: "A",
         scope: false,
         replace: false,
-        transclude: false,
         require: 'leaflet',
 
-        link: function($scope, element, attrs, controller) {
+        link: function(scope, element, attrs, controller) {
+            var isDefined = leafletHelpers.isDefined,
+                createLeafletBounds = leafletHelpers.createLeafletBounds,
+                leafletScope = controller.getLeafletScope();
+
             controller.getMap().then(function(map) {
-                setupBounds(map);
 
-                function isBoundsValid(bounds) {
-                    return isDefined(bounds) && isDefined(bounds.southWest) &&
-                        isDefined(bounds.northEast) && isNumber(bounds.southWest.lat) &&
-                        isNumber(bounds.southWest.lng) && isNumber(bounds.northEast.lat) &&
-                        isNumber(bounds.northEast.lng);
+                function updateBoundsInScope() {
+                    if(!leafletScope.bounds) { return; }
+
+                    var mapBounds = map.getBounds();
+                    var newScopeBounds = {
+                        northEast: {
+                            lat: mapBounds.getNorthEast().lat,
+                            lng: mapBounds.getNorthEast().lng
+                        },
+                        southWest: {
+                            lat: mapBounds.getSouthWest().lat,
+                            lng: mapBounds.getSouthWest().lng
+                        }
+                    };
+
+                    if(!angular.equals(leafletScope.bounds, newScopeBounds)) {
+                        leafletScope.bounds = newScopeBounds;
+                    }
                 }
 
-                function setupBounds(map) {
-                    $scope.$watch('bounds', function(bounds) {
-                        if (!isDefined(bounds) || !isBoundsValid(bounds)) {
-                            $log.error('[AngularJS - Leaflet] Invalid bounds');
-                            return;
-                        }
+                function boundsListener(newBounds) {
+                    if (!isDefined(newBounds)) {
+                        $log.error('[AngularJS - Leaflet] Invalid bounds');
+                        return;
+                    }
 
-                        var southWest = bounds.southWest;
-                        var northEast = bounds.northEast;
-                        var new_latlng_bounds = new L.LatLngBounds(
-                                new L.LatLng(southWest.lat, southWest.lng),
-                                new L.LatLng(northEast.lat, northEast.lng));
-
-                        if (!map.getBounds().equals(new_latlng_bounds)) {
-                            map.fitBounds(new_latlng_bounds);
-                        }
-                    }, true);
+                    var leafletBounds = createLeafletBounds(newBounds);
+                    if (leafletBounds && !map.getBounds().equals(leafletBounds)) {
+                        map.fitBounds(leafletBounds);
+                    }
                 }
+
+                map.on('moveend', updateBoundsInScope);
+                map.on('dragend', updateBoundsInScope);
+                map.on('zoomend', updateBoundsInScope);
+
+                map.whenReady(function() {
+                    leafletScope.$watch('bounds', boundsListener, true);
+                });
             });
         }
     };