+++ src/directives/controls.js
@@ -1,16 +1,18 @@
-angular.module("leaflet-directive").directive('controls', function ($log) {
+angular.module("leaflet-directive").directive('controls', function ($log, leafletHelpers) {
     return {
         restrict: "A",
         scope: false,
         replace: false,
-        transclude: false,
         require: 'leaflet',
 
-        link: function($scope, element, attrs, controller) {
-            var controls = $scope.controls;
+        link: function(scope, element, attrs, controller) {
+            var isDefined = leafletHelpers.isDefined,
+                leafletScope  = controller.getLeafletScope(),
+                controls = leafletScope.controls;
+
             controller.getMap().then(function(map) {
-                if (isDefined(controls.draw)) {
-                    var drawControl = new L.Control.Draw();
+                if (isDefined(L.Control.Draw) && isDefined(controls.draw)) {
+                    var drawControl = new L.Control.Draw(controls.draw.options);
                     map.addControl(drawControl);
                 }
             });