+++ src/directives/eventBroadcast.js
@@ -1,110 +1,53 @@
-angular.module("leaflet-directive").directive('eventBroadcast', function ($log, $rootScope) {
+angular.module("leaflet-directive").directive('eventBroadcast', function ($log, $rootScope, leafletHelpers, leafletEvents) {
     return {
         restrict: "A",
         scope: false,
         replace: false,
-        transclude: false,
         require: 'leaflet',
 
-        link: function($scope, element, attrs, controller) {
+        link: function(scope, element, attrs, controller) {
+            var isObject = leafletHelpers.isObject,
+                leafletScope  = controller.getLeafletScope(),
+                eventBroadcast = leafletScope.eventBroadcast,
+                availableMapEvents = leafletEvents.getAvailableMapEvents(),
+                genDispatchMapEvent = leafletEvents.genDispatchMapEvent;
+
             controller.getMap().then(function(map) {
-
-                function genDispatchMapEvent(eventName, logic) {
-                    return function(e) {
-                        // Put together broadcast name
-                        var broadcastName = 'leafletDirectiveMap.' + eventName;
-                        // Safely broadcast the event
-                        safeApply($scope, function(scope) {
-                            if (logic === "emit") {
-                                scope.$emit(broadcastName, {
-                                    leafletEvent : e
-                                });
-                            } else if (logic === "broadcast") {
-                                $rootScope.$broadcast(broadcastName, {
-                                    leafletEvent : e
-                                });
-                            }
-                        });
-                    };
-                }
-
-                var availableMapEvents = [
-                    'click',
-                    'dblclick',
-                    'mousedown',
-                    'mouseup',
-                    'mouseover',
-                    'mouseout',
-                    'mousemove',
-                    'contextmenu',
-                    'focus',
-                    'blur',
-                    'preclick',
-                    'load',
-                    'unload',
-                    'viewreset',
-                    'movestart',
-                    'move',
-                    'moveend',
-                    'dragstart',
-                    'drag',
-                    'dragend',
-                    'zoomstart',
-                    'zoomend',
-                    'zoomlevelschange',
-                    'resize',
-                    'autopanstart',
-                    'layeradd',
-                    'layerremove',
-                    'baselayerchange',
-                    'overlayadd',
-                    'overlayremove',
-                    'locationfound',
-                    'locationerror',
-                    'popupopen',
-                    'popupclose'
-                ];
 
                 var mapEvents = [];
                 var i;
                 var eventName;
                 var logic = "broadcast";
 
-                if (!isDefinedAndNotNull($scope.eventBroadcast)) {
-                    // Backward compatibility, if no event-broadcast attribute, all events are broadcasted
-                    mapEvents = availableMapEvents;
-                } else if (typeof $scope.eventBroadcast !== 'object') {
-                    // Not a valid object
-                    $log.warn("[AngularJS - Leaflet] event-broadcast must be an object check your model.");
-                } else {
+                if (isObject(eventBroadcast)) {
                     // We have a possible valid object
-                    if ($scope.eventBroadcast.map === undefined || $scope.eventBroadcast.map === null) {
+                    if (eventBroadcast.map === undefined || eventBroadcast.map === null) {
                         // We do not have events enable/disable do we do nothing (all enabled by default)
                         mapEvents = availableMapEvents;
-                    } else if (typeof $scope.eventBroadcast.map !== 'object') {
+                    } else if (typeof eventBroadcast.map !== 'object') {
                         // Not a valid object
                         $log.warn("[AngularJS - Leaflet] event-broadcast.map must be an object check your model.");
                     } else {
                         // We have a possible valid map object
                         // Event propadation logic
-                        if ($scope.eventBroadcast.map.logic !== undefined && $scope.eventBroadcast.map.logic !== null) {
+                        if (eventBroadcast.map.logic !== undefined && eventBroadcast.map.logic !== null) {
                             // We take care of possible propagation logic
-                            if ($scope.eventBroadcast.map.logic !== "emit" && $scope.eventBroadcast.map.logic !== "broadcast") {
+                            if (eventBroadcast.map.logic !== "emit" && eventBroadcast.map.logic !== "broadcast") {
                                 // This is an error
                                 $log.warn("[AngularJS - Leaflet] Available event propagation logic are: 'emit' or 'broadcast'.");
-                            } else if ($scope.eventBroadcast.map.logic === "emit") {
+                            } else if (eventBroadcast.map.logic === "emit") {
                                 logic = "emit";
                             }
                         }
                         // Enable / Disable
                         var mapEventsEnable = false, mapEventsDisable = false;
-                        if ($scope.eventBroadcast.map.enable !== undefined && $scope.eventBroadcast.map.enable !== null) {
-                            if (typeof $scope.eventBroadcast.map.enable === 'object') {
+                        if (eventBroadcast.map.enable !== undefined && eventBroadcast.map.enable !== null) {
+                            if (typeof eventBroadcast.map.enable === 'object') {
                                 mapEventsEnable = true;
                             }
                         }
-                        if ($scope.eventBroadcast.map.disable !== undefined && $scope.eventBroadcast.map.disable !== null) {
-                            if (typeof $scope.eventBroadcast.map.disable === 'object') {
+                        if (eventBroadcast.map.disable !== undefined && eventBroadcast.map.disable !== null) {
+                            if (typeof eventBroadcast.map.disable === 'object') {
                                 mapEventsDisable = true;
                             }
                         }
@@ -118,8 +61,8 @@ angular.module("leaflet-directive").directive('eventBroadcast', function ($log, 
                             // At this point the map object is OK, lets enable or disable events
                             if (mapEventsEnable) {
                                 // Enable events
-                                for (i = 0; i < $scope.eventBroadcast.map.enable.length; i++) {
-                                    eventName = $scope.eventBroadcast.map.enable[i];
+                                for (i = 0; i < eventBroadcast.map.enable.length; i++) {
+                                    eventName = eventBroadcast.map.enable[i];
                                     // Do we have already the event enabled?
                                     if (mapEvents.indexOf(eventName) !== -1) {
                                         // Repeated event, this is an error
@@ -138,8 +81,8 @@ angular.module("leaflet-directive").directive('eventBroadcast', function ($log, 
                             } else {
                                 // Disable events
                                 mapEvents = availableMapEvents;
-                                for (i = 0; i < $scope.eventBroadcast.map.disable.length; i++) {
-                                    eventName = $scope.eventBroadcast.map.disable[i];
+                                for (i = 0; i < eventBroadcast.map.disable.length; i++) {
+                                    eventName = eventBroadcast.map.disable[i];
                                     var index = mapEvents.indexOf(eventName);
                                     if (index === -1) {
                                         // The event does not exist
@@ -151,13 +94,16 @@ angular.module("leaflet-directive").directive('eventBroadcast', function ($log, 
                             }
                         }
                     }
-                }
 
-                for (i = 0; i < mapEvents.length; i++) {
-                    eventName = mapEvents[i];
-                    map.on(eventName, genDispatchMapEvent(eventName, logic), {
-                        eventName: eventName
-                    });
+                    for (i = 0; i < mapEvents.length; i++) {
+                        eventName = mapEvents[i];
+                        map.on(eventName, genDispatchMapEvent(leafletScope, eventName, logic), {
+                            eventName: eventName
+                        });
+                    }
+                } else {
+                    // Not a valid object
+                    $log.warn("[AngularJS - Leaflet] event-broadcast must be an object, check your model.");
                 }
             });
         }