+++ src/directives/leaflet.js
@@ -1,8 +1,8 @@
-angular.module("leaflet-directive", []).directive('leaflet', function ($log, $q, leafletData) {
+angular.module("leaflet-directive", []).directive('leaflet', function ($log, $q, leafletData, leafletMapDefaults, leafletHelpers, leafletEvents) {
+    var _leafletMap;
     return {
-        restrict: "E",
+        restrict: "EA",
         replace: true,
-        transclude: true,
         scope: {
             center: '=center',
             defaults: '=defaults',
@@ -16,22 +16,29 @@ angular.module("leaflet-directive", []).directive('leaflet', function ($log, $q,
             tiles: '=tiles',
             layers: '=layers',
             controls: '=controls',
-            eventBroadcast: '=eventBroadcast'
+            eventBroadcast: '=eventBroadcast',
+            watchMarkers: '=watchmarkers'
         },
-        template: '<div class="angular-leaflet-map" ng-transclude></div>',
+        template: '<div class="angular-leaflet-map"></div>',
         controller: function ($scope) {
-            $scope.leafletMap = $q.defer();
+            _leafletMap = $q.defer();
             this.getMap = function () {
-                return $scope.leafletMap.promise;
+                return _leafletMap.promise;
+            };
+
+            this.getLeafletScope = function() {
+                return $scope;
             };
         },
 
-        link: function($scope, element, attrs, controller) {
-            var defaults = parseMapDefaults($scope.defaults);
-            leafletData.setDefaults(defaults);
+        link: function(scope, element, attrs) {
+            var isDefined = leafletHelpers.isDefined,
+                defaults = leafletMapDefaults.setDefaults(scope.defaults, attrs.id),
+                genDispatchMapEvent = leafletEvents.genDispatchMapEvent,
+                mapEvents = leafletEvents.getAvailableMapEvents();
 
             // If we are going to set maxBounds, undefine the minZoom property
-            if (isDefined($scope.maxBounds)) {
+            if (isDefined(scope.maxBounds)) {
                 defaults.minZoom = undefined;
             }
 
@@ -51,31 +58,17 @@ angular.module("leaflet-directive", []).directive('leaflet', function ($log, $q,
                 }
             }
 
-            if (isDefined(attrs.marker)) {
-                 $log.warn("[AngularJS - Leaflet] The 'marker' property is currently deprecated, please use the 'markers' property instead.");
-            }
-
             // Create the Leaflet Map Object with the options
-            var map = new L.Map(element[0], {
-                maxZoom: defaults.maxZoom,
-                minZoom: defaults.minZoom,
-                keyboard: defaults.keyboard,
-                dragging: defaults.dragging,
-                zoomControl: defaults.zoomControl,
-                doubleClickZoom: defaults.doubleClickZoom,
-                scrollWheelZoom: defaults.scrollWheelZoom,
-                attributionControl: defaults.attributionControl
-            });
-
-            $scope.leafletMap.resolve(map);
-            leafletData.setMap(map, attrs.id);
+            var map = new L.Map(element[0], leafletMapDefaults.getMapCreationDefaults(attrs.id));
+            _leafletMap.resolve(map);
 
             if (!isDefined(attrs.center)) {
-                 $log.warn("[AngularJS - Leaflet] 'center' is undefined in the current scope, did you forget to initialize it?");
-                 map.setView([defaults.center.lat, defaults.center.lng], defaults.center.zoom);
+                $log.warn("[AngularJS - Leaflet] 'center' is undefined in the current scope, did you forget to initialize it?");
+                map.setView([defaults.center.lat, defaults.center.lng], defaults.center.zoom);
             }
 
-            if (!isDefined(attrs.tiles) && !isDefined(attrs.layers)) {
+            // If no layers nor tiles defined, set the default tileLayer
+            if (!isDefined(attrs.tiles) && (!isDefined(attrs.layers))) {
                 var tileLayerObj = L.tileLayer(defaults.tileLayer, defaults.tileLayerOptions);
                 tileLayerObj.addTo(map);
                 leafletData.setTiles(tileLayerObj);
@@ -85,6 +78,22 @@ angular.module("leaflet-directive", []).directive('leaflet', function ($log, $q,
             if (isDefined(map.zoomControl) && isDefined(defaults.zoomControlPosition)) {
                 map.zoomControl.setPosition(defaults.zoomControlPosition);
             }
+
+            // if no event-broadcast attribute, all events are broadcasted
+            if (!isDefined(attrs.eventBroadcast)) {
+                var logic = "broadcast";
+                for (var i = 0; i < mapEvents.length; i++) {
+                    var eventName = mapEvents[i];
+                    map.on(eventName, genDispatchMapEvent(scope, eventName, logic), {
+                        eventName: eventName
+                    });
+                }
+            }
+
+            // Resolve the map object to the promises
+            map.whenReady(function() {
+                leafletData.setMap(map, attrs.id);
+            });
         }
     };
 });