+++ src/directives/legend.js
@@ -1,22 +1,23 @@
-angular.module("leaflet-directive").directive('legend', function ($log) {
+angular.module("leaflet-directive").directive('legend', function ($log, leafletHelpers) {
     return {
         restrict: "A",
         scope: false,
         replace: false,
-        transclude: false,
         require: 'leaflet',
 
-        link: function($scope, element, attrs, controller) {
-            controller.getMap().then(function(map) {
-                var legend = $scope.legend;
+        link: function(scope, element, attrs, controller) {
+            var isArray      = leafletHelpers.isArray,
+                leafletScope = controller.getLeafletScope(),
+                legend       = leafletScope.legend;
 
+            controller.getMap().then(function(map) {
                 if (!isArray(legend.colors) || !isArray(legend.labels) || legend.colors.length !== legend.labels.length) {
                     $log.warn("[AngularJS - Leaflet] legend.colors and legend.labels must be set.");
                 } else {
                     var legendClass = legend.legendClass ? legend.legendClass : "legend";
                     var position = legend.position || 'bottomright';
                     var leafletLegend = L.control({ position: position });
-                    leafletLegend.onAdd = function (map) {
+                    leafletLegend.onAdd = function (/*map*/) {
                         var div = L.DomUtil.create('div', legendClass);
                         for (var i = 0; i < legend.colors.length; i++) {
                             div.innerHTML +=