+++ src/directives/maxBounds.js
@@ -1,29 +1,30 @@
-angular.module("leaflet-directive").directive('maxbounds', function ($log) {
+angular.module("leaflet-directive").directive('maxbounds', function ($log, leafletMapDefaults, leafletHelpers) {
     return {
         restrict: "A",
         scope: false,
         replace: false,
-        transclude: false,
         require: 'leaflet',
 
-        link: function($scope, element, attrs, controller) {
-            var defaults = parseMapDefaults($scope.defaults);
+        link: function(scope, element, attrs, controller) {
+            var leafletScope  = controller.getLeafletScope(),
+                isValidBounds = leafletHelpers.isValidBounds;
+
+
             controller.getMap().then(function(map) {
-            var maxBounds = $scope.maxBounds;
-                if (isDefined(maxBounds) && isDefined(maxBounds.southWest) && isDefined(maxBounds.northEast)) {
-                    $scope.$watch("maxBounds", function (maxBounds) {
-                        if (isDefined(maxBounds.southWest) && isDefined(maxBounds.northEast) && isNumber(maxBounds.southWest.lat) && isNumber(maxBounds.southWest.lng) && isNumber(maxBounds.northEast.lat) && isNumber(maxBounds.northEast.lng)) {
-                            map.setMaxBounds(
-                                new L.LatLngBounds(
-                                    new L.LatLng(maxBounds.southWest.lat, maxBounds.southWest.lng),
-                                    new L.LatLng(maxBounds.northEast.lat, maxBounds.northEast.lng)
-                                )
-                            );
-                        }
-                    });
-                } else {
-                    $log.warn("[AngularJS - Leaflet] 'maxBounds' is defined in the current scope, but not correctly initialized.");
-                }
+                leafletScope.$watch("maxBounds", function (maxBounds) {
+                    if (!isValidBounds(maxBounds)) {
+                        // Unset any previous maxbounds
+                        map.setMaxBounds();
+                        return;
+                    }
+                    map.setMaxBounds(
+                        new L.LatLngBounds(
+                            new L.LatLng(maxBounds.southWest.lat, maxBounds.southWest.lng),
+                            new L.LatLng(maxBounds.northEast.lat, maxBounds.northEast.lng)
+                        ),
+                        maxBounds.options
+                    );
+                });
             });
         }
     };