+++ test/unit/leafletDirectiveSpec.js
@@ -5,14 +5,15 @@
 /* jasmine specs for directives go here */
 
 describe('Directive: leaflet', function() {
-    var $compile = null, $rootScope = null, $timeout, leafletData = null;
+    var $compile = null, $rootScope = null, $timeout, leafletData = null, leafletMapDefaults = null;
 
     beforeEach(module('leaflet-directive'));
-    beforeEach(inject(function(_$compile_, _$rootScope_, _$timeout_, _leafletData_){
+    beforeEach(inject(function(_$compile_, _$rootScope_, _$timeout_, _leafletData_, _leafletMapDefaults_) {
         $compile = _$compile_;
         $rootScope = _$rootScope_;
         $timeout = _$timeout_;
         leafletData = _leafletData_;
+        leafletMapDefaults = _leafletMapDefaults_;
     }));
 
     afterEach(inject(function($rootScope) {
@@ -22,43 +23,57 @@ describe('Directive: leaflet', function() {
     it('should have loaded leaflet library inside the directive', function() {
         var element = angular.element('<leaflet></leaflet>');
         element = $compile(element)($rootScope);
+        $rootScope.$digest();
         expect(element.text()).toEqual('+-Leaflet | © OpenStreetMap contributors');
     });
 
     it('should set default center if not center is provided', function() {
         var element = angular.element('<leaflet></leaflet>');
         element = $compile(element)($rootScope);
-        leafletData.getMap($rootScope).then(function(map) {
-            expect(map.getZoom()).toEqual(1);
-            expect(map.getCenter().lat).toEqual(0);
-            expect(map.getCenter().lng).toEqual(0);
+        var leafletMap;
+        leafletData.getMap().then(function(map) {
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.getZoom()).toEqual(1);
+        expect(leafletMap.getCenter().lat).toEqual(0);
+        expect(leafletMap.getCenter().lng).toEqual(0);
     });
 
     it('should set default tile if not tiles nor layers are provided', function() {
         var element = angular.element('<leaflet></leaflet>');
         element = $compile(element)($rootScope);
-        leafletData.getTiles().then(function(leafletTiles) {
-            expect(leafletTiles._url).toEqual(leafletData.getDefaults().tileLayer);
+        var leafletTiles;
+        leafletData.getTiles().then(function(tiles) {
+            leafletTiles = tiles;
         });
+        $rootScope.$digest();
+        var defaults = leafletMapDefaults.getDefaults();
+        expect(leafletTiles._url).toEqual(defaults.tileLayer);
     });
 
     it('should set the max zoom if specified', function() {
         angular.extend($rootScope, { defaults: { maxZoom: 15 } });
         var element = angular.element('<leaflet defaults="defaults"></leaflet>');
         element = $compile(element)($rootScope);
+        var leafletMap;
         leafletData.getMap().then(function(map) {
-            expect(map.getMaxZoom()).toEqual(15);
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.getMaxZoom()).toEqual(15);
     });
 
     it('should set the min zoom if specified', function() {
         angular.extend($rootScope, { defaults: { minZoom: 4 } });
         var element = angular.element('<leaflet defaults="defaults"></leaflet>');
         element = $compile(element)($rootScope);
+        var leafletMap;
         leafletData.getMap().then(function(map) {
-            expect(map.getMinZoom()).toEqual(4);
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.getMinZoom()).toEqual(4);
     });
 
     it('should unset the minzoom if maxbounds specified', function() {
@@ -79,9 +94,12 @@ describe('Directive: leaflet', function() {
         });
         var element = angular.element('<leaflet defaults="defaults" maxBounds="maxBounds"></leaflet>');
         element = $compile(element)($rootScope);
+        var leafletMap;
         leafletData.getMap().then(function(map) {
-            expect(map.getMinZoom()).toEqual(0);
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.getMinZoom()).toEqual(0);
     });
 
     it('should set tileLayer and tileLayer options if specified', function() {
@@ -96,12 +114,16 @@ describe('Directive: leaflet', function() {
         });
         var element = angular.element('<leaflet defaults="defaults"></leaflet>');
         element = $compile(element)($rootScope);
-        leafletData.getTiles().then(function(leafletTiles) {
-            expect(leafletTiles.options.detectRetina).toEqual(true);
-            expect(leafletTiles.options.opacity).toEqual(0.8);
-            expect(leafletTiles._url).toEqual("http://{s}.tile.opencyclemap.org/cycle/{z}/{x}/{y}.png");
-            expect(leafletData.getDefaults().tileLayer).toEqual("http://{s}.tile.opencyclemap.org/cycle/{z}/{x}/{y}.png");
+        var leafletTiles;
+        leafletData.getTiles().then(function(tiles) {
+            leafletTiles = tiles;
         });
+        $rootScope.$digest();
+        var defaults = leafletMapDefaults.getDefaults();
+        expect(leafletTiles.options.detectRetina).toEqual(true);
+        expect(leafletTiles.options.opacity).toEqual(0.8);
+        expect(leafletTiles._url).toEqual("http://{s}.tile.opencyclemap.org/cycle/{z}/{x}/{y}.png");
+        expect(defaults.tileLayer).toEqual("http://{s}.tile.opencyclemap.org/cycle/{z}/{x}/{y}.png");
     });
 
     it('should set zoom control button properly if zoomControlPosition option is set', function() {
@@ -112,9 +134,12 @@ describe('Directive: leaflet', function() {
         });
         var element = angular.element('<leaflet defaults="defaults"></leaflet>');
         element = $compile(element)($rootScope);
+        var leafletMap;
         leafletData.getMap().then(function(map) {
-            expect(map.zoomControl.getPosition()).toEqual('topright');
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.zoomControl.getPosition()).toEqual('topright');
     });
 
     it('should remove zoom control button if unset on defaults', function() {
@@ -125,8 +150,11 @@ describe('Directive: leaflet', function() {
         });
         var element = angular.element('<leaflet defaults="defaults"></leaflet>');
         element = $compile(element)($rootScope);
+        var leafletMap;
         leafletData.getMap().then(function(map) {
-            expect(map.zoomControl).toBe(undefined);
+            leafletMap = map;
         });
+        $rootScope.$digest();
+        expect(leafletMap.zoomControl).toBe(undefined);
     });
 });